requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
paho-mqtt==2.1.0
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
python-multipart==0.0.12
